Update script create Azure resources through ARM. Also updated EventCount to work with Kafka

+++ EventCountExample/EventCountTopology/src/main/java/com/microsoft/hdinsight/storm/examples/DBGlobalCountBolt.java
@@ -18,6 +18,7 @@
 package com.microsoft.hdinsight.storm.examples;
 
 import java.util.Map;
+import java.util.LinkedList;
 
 import backtype.storm.Config;
 import backtype.storm.Constants;
@@ -25,37 +26,63 @@ import backtype.storm.Constants;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import backtype.storm.task.OutputCollector;
 import backtype.storm.task.TopologyContext;
-import backtype.storm.topology.BasicOutputCollector;
 import backtype.storm.topology.OutputFieldsDeclarer;
-import backtype.storm.topology.base.BaseBasicBolt;
+import backtype.storm.topology.base.BaseRichBolt;
 import backtype.storm.tuple.Tuple;
+import backtype.storm.tuple.Values;
 
 /**
  * Globally count number of messages
  */
-public class DBGlobalCountBolt extends BaseBasicBolt {
+public class DBGlobalCountBolt extends BaseRichBolt {
   private static final long serialVersionUID = 1L;
   private static final Logger logger = LoggerFactory
       .getLogger(DBGlobalCountBolt.class);
 
-  private long partialCount = 0L;
+  private long finalCount = 0L;
   private long totalCount = 0L;
+  
+  private LinkedList<Tuple> tuplesToAck = new LinkedList<Tuple>();
   
   private SqlDb db;
   private String connectionStr;
+  private String tableName = "EventCount";
   
-  public DBGlobalCountBolt(String connectionStr) {
+  private int tickTupleFreqSecs = 5;
+  
+  private OutputCollector collector;
+  
+  Boolean enableAck = false;
+  
+  public DBGlobalCountBolt(Boolean enableAck, String connectionStr, String table) {
+    this(enableAck, connectionStr, table, 5);
+  }
+  
+  public DBGlobalCountBolt(Boolean enableAck, String connectionStr, String table, int tickTupleFreq) {
+    this.enableAck = enableAck;
+    logger.info("enableAck = " + enableAck);
+    this.tickTupleFreqSecs = tickTupleFreq;
+    logger.info("tickTupleFreqSecs = " + tickTupleFreqSecs);
+    
     this.connectionStr = connectionStr;
+    logger.info("connectionStr = " + this.connectionStr);
+    this.tableName = table;
+    logger.info("tableName = " + this.tableName);
   }
   
   @Override
-  public void prepare(Map config, TopologyContext context) {
-    db = new SqlDb(connectionStr);
-    db.dropTable();
-    db.createTable();
-    partialCount = 0L;
+  public void prepare(Map config, TopologyContext context, OutputCollector collector) {
+    if(connectionStr != null && !connectionStr.isEmpty()) {
+      logger.info("prepare - connectionStr = " + connectionStr + ", tableName = " + this.tableName);
+      db = new SqlDb(connectionStr, tableName);
+      db.dropTable();
+      db.createTable();
+    }
+    finalCount = 0L;
     totalCount = 0L;
+    this.collector = collector;
   }
   
   @Override
@@ -64,24 +91,45 @@ public class DBGlobalCountBolt extends BaseBasicBolt {
   }
 
   @Override
-  public void execute(Tuple tuple, BasicOutputCollector collector) {
-    //write the partialCount to SqlDb on each TickTuple.
+  public void execute(Tuple tuple) {
+    //write the finalCount to SqlDb on each TickTuple.
     if (isTickTuple(tuple)) {
-      //only emit if partialCount > 0
-      if(partialCount > 0) {
-        logger.info("updating database" + 
-            ", partialCount: " + partialCount + 
-            ", totalCount: " + totalCount);
-        db.insertValue(System.currentTimeMillis(), partialCount);
-        partialCount = 0L;
+      //only emit if finalCount > 0
+      if(finalCount > 0) {
+        logger.info("emitting final count" + 
+            ", finalCount: " + finalCount + 
+            ", totalCount: " + totalCount + 
+            ", tuplesToAck: " + tuplesToAck.size() + 
+            " at " + System.currentTimeMillis());
+        if(db != null) {
+          db.insertValue(System.currentTimeMillis(), finalCount);
+        } else {
+          if(enableAck) {
+            this.collector.emit(tuplesToAck, new Values(finalCount));
+          } else {
+            this.collector.emit(new Values(finalCount));
+          }
+        }
+        
+        finalCount = 0L;
+        if(enableAck) {
+          for(Tuple tupleToAck : tuplesToAck)
+          {
+            collector.ack(tupleToAck);
+          }
+          tuplesToAck.clear();
+        }
       }
     }
     else
     {
-      //Merge partialCount from all PartialCountBolt tasks
+      //Merge finalCount from all PartialCountBolt tasks
       long incomingPartialCount = tuple.getLong(0);
-      partialCount += incomingPartialCount;
+      finalCount += incomingPartialCount;
       totalCount += incomingPartialCount;
+      if(enableAck) {
+        tuplesToAck.add(tuple);
+      }
     }
   }
 
@@ -92,8 +140,8 @@ public class DBGlobalCountBolt extends BaseBasicBolt {
   @Override
   public Map<String,Object>  getComponentConfiguration() {
     Config conf = new Config();
-    //set the TickTuple frequency to 1 second
-    conf.put(Config.TOPOLOGY_TICK_TUPLE_FREQ_SECS, 1);
+    //set the TickTuple frequency
+    conf.put(Config.TOPOLOGY_TICK_TUPLE_FREQ_SECS, tickTupleFreqSecs);
     return conf;
   }
